bpf: fix plugins not stopping when context cancelled
When the client event table changes, Velociraptor cancels each plugin's
context and expects it to shut itself down. Then it starts them again
if they are still needed in the new table. The bpf based plugins were
not stopping and Velociraptor was starting new goroutines for them,
leaving the old ones running and leaking resources forever.

This commit changes the bpf plugins so they now exit when the context
is cancelled. Velociraptor does not wait until the plugins are fully
stopped before starting new instances and the concurrent loading and
unloading of BPF modules is problematic. To ensure only one BPF module
load or unload operation is happening at any one time, this commit
introduces a publish/subscribe service.